# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "private project Api Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "namsookill@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.dripntrip.test/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/user": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "새로운 유저 생성",
                "description": "회원 가입을 위한 API 입니다. 사용자 아이디, 이름, 암호, 암호 확인을 입력해야 합니다.",
                "operationId": "user",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "유저 생성"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "USER"
                ],
                "summary": "새로운 유저 생성",
                "description": "회원 로그인 위한 API 입니다. 사용자 아이디, 암호를 입력해야 합니다.",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "유저 로그인 토큰 발급"
                    },
                    "422": {
                        "description": "HTTP_UNPROCESSABLE_ENTITY"
                    },
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "USER"
                ],
                "summary": "유저 로그아웃",
                "description": "회원 로그인 아웃을 위한 API 입니다. 토큰 만료",
                "operationId": "logout",
                "responses": {
                    "500": {
                        "description": "HTTP_INTERNAL_SERVER_ERROR"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {}
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "USER",
            "description": "API Endpoints of user"
        }
    ]
}